fix: Add .nvmrc to allow users to deploy on cloudflare pages (#269)
This project can be hosted on cf pages, as the code in gh actions. The
problem is that the node version used in cf pages is not 20 which will
cause the deployment failing.

This PR adds a .nvmrc that specifies node version to 20, so cf pages can
be successfully deployed.

To deploy it:

1. Fork this repo
2. In cf pages, create a new project from forked repo
3. Choose framework "SvelteKit"
4. Click deploy

---------

Co-authored-by: Fernando Maclen <fmaclen@gmail.com>